Implement JVM_GetProperties for jdk17+
This reverts commit f12ef363bdf96e4846ff4276fa3790c59b4f4925, which
re-instates commit 9620a5a7654118a6b379a8185281653d06b3fa80, PR
https://github.com/eclipse-openj9/openj9/pull/21200 plus some additional
cosmetic changes, and the following.

Setting SystemProps.customTmpdir is unnecessary, and fails.
It already has the correct value from calling
SystemProps.initProperties().

Issue https://github.com/eclipse-openj9/openj9/issues/21311

Modify the tracepoint used by test to match VM changes.
cmdLineTester_PageAlignDirectMemory uses the j9jcl.352 tracepoint, which
has moved to j9vm.791.

Issue https://github.com/eclipse-openj9/openj9/issues/21314

Update JVM_GetTemporaryDirectory
