+++ runtime/jcl/common/system.c
@@ -44,27 +44,7 @@
 #endif
 
 
-
-/* JCL_J2SE */
-#define JCL_J2SE
-
-
-
-typedef struct {
-	int errorOccurred;
-	jobject args;
-	jint nCommandLineDefines;
-	JNIEnv *env;
-	const char **defaultValues;
-	int defaultCount;
-	jclass stringClass;
-} CreateSystemPropertiesData;
-
-jint propertyListAddString( JNIEnv *env, jarray array, jint arrayIndex, const char *value);
-static void JNICALL systemPropertyIterator(char* key, char* value, void* userData);
-jobject getPropertyList(JNIEnv *env);
-
-#if JAVA_SPEC_VERSION >= 11
+#if JAVA_SPEC_VERSION == 11
 void JNICALL
 Java_java_lang_System_initJCLPlatformEncoding(JNIEnv *env, jclass clazz)
 {
@@ -93,7 +73,7 @@ Java_java_lang_System_initJCLPlatformEncoding(JNIEnv *env, jclass clazz)
 		}
 	}
 }
-#endif /* JAVA_SPEC_VERSION >= 11 */
+#endif /* JAVA_SPEC_VERSION == 11 */
 
 /**
  * sysPropID
@@ -172,12 +152,8 @@ Java_java_lang_System_getSysPropBeforePropertiesInitialized(JNIEnv *env, jclass 
 		}
 		break;
 
-	case 4: /* default value of java.io.tmpDir before any -D options */
-		sysPropValue = getTmpDir(env, &envSpace);
-		break;
-
 #if defined(J9ZOS390) && (JAVA_SPEC_VERSION >= 21)
-	case 5: /* com.ibm.autocvt setting on z/OS */
+	case 4: /* com.ibm.autocvt setting on z/OS */
 		sysPropValue = getDefinedArgumentFromJavaVMInitArgs(vmInitArgs, "com.ibm.autocvt");
 		if (NULL == sysPropValue) {
 			/* As part of better handling of JEP400 constraints on z/OS, the com.ibm.autocvt property
@@ -203,10 +179,13 @@ Java_java_lang_System_getSysPropBeforePropertiesInitialized(JNIEnv *env, jclass 
 	return result;
 }
 
-jobject JNICALL Java_java_lang_System_getPropertyList(JNIEnv *env, jclass clazz)
+#if JAVA_SPEC_VERSION < 17
+jobjectArray JNICALL
+Java_java_lang_System_getPropertyList(JNIEnv *env, jclass clazz)
 {
-	return getPropertyList(env);
+	return ((J9VMThread *)env)->javaVM->internalVMFunctions->getSystemPropertyList(env);
 }
+#endif /* JAVA_SPEC_VERSION < 17 */
 
 jstring JNICALL Java_java_lang_System_mapLibraryName(JNIEnv * env, jclass unusedClass, jstring inName)
 {
@@ -258,350 +237,6 @@ void JNICALL Java_java_lang_System_setFieldImpl(JNIEnv * env, jclass cls, jstrin
 	if (!descriptorFID) return;
 	(*env)->SetStaticObjectField(env, cls, descriptorFID, stream);
 }
-
-
-jobject createSystemPropertyList(JNIEnv *env, const char *defaultValues[], int defaultCount)
-{
-	VMI_ACCESS_FROM_ENV(env);
-
-	jint i, nCommandLineDefines = 0;
-	jclass stringClass;
-	jarray args = NULL;
-	int propertyCount;
-
-	stringClass = (*env)->FindClass(env, "java/lang/String");
-	if (!stringClass) {
-/*		printf("\nFailed to find class java/lang/String");*/
-		return (jobject) 0;
-	}
-
-	(*VMI)->CountSystemProperties(VMI, &propertyCount);
-	if (propertyCount) {
-		CreateSystemPropertiesData iteratorData;
-
-		args = (*env)->NewObjectArray(env, defaultCount + (propertyCount * 2), stringClass, NULL);
-		if (NULL == args) {
-			return NULL;
-		}
-
-		iteratorData.errorOccurred = 0;
-		iteratorData.args = args;
-		iteratorData.nCommandLineDefines = nCommandLineDefines;
-		iteratorData.env = env;
-		iteratorData.defaultValues = defaultValues;
-		iteratorData.defaultCount = defaultCount;
-
-		iteratorData.stringClass = stringClass;
-		(*VMI)->IterateSystemProperties(VMI, systemPropertyIterator, &iteratorData);
-		if (iteratorData.errorOccurred) {
-			return NULL;
-		}
-		nCommandLineDefines = iteratorData.nCommandLineDefines;
-	}
-
-	if (NULL == args) {
-		args = (*env)->NewObjectArray(env, defaultCount, stringClass, NULL);
-	}
-	if (NULL == args) {
-/*		printf("\nFailed to create arg array");*/
-		return NULL;
-	}
-
-	for (i = 0; i < defaultCount; ++i) {
-		if (defaultValues[i] == NULL) continue;
-		if (-1 == propertyListAddString( env, args, nCommandLineDefines, defaultValues[i]) ) {
-			return NULL;
-		}
-		nCommandLineDefines++;
-	}
-
-	return args;
-}
-
-/**
- * @return 0 on success, -1 on error
- */
-jint propertyListAddString( JNIEnv *env, jarray array, jint arrayIndex, const char *value)
-{
-	/* String must be well-formed modified UTF-8 */
-	jobject str = (*env)->NewStringUTF(env, value);
-	if (NULL != str) {
-		(*env)->SetObjectArrayElement(env, array, arrayIndex, str);
-	}
-	/* NewStringUTF does not throw an exception, other than OutOfMemory */
-	return ((*env)->ExceptionCheck(env) == JNI_TRUE) ? -1 : 0;
-}
-
-jobject getPropertyList(JNIEnv *env)
-{
-	PORT_ACCESS_FROM_ENV(env);
-	int propIndex = 0;
-	jobject propertyList = NULL;
-#define PROPERTY_COUNT 137
-	char *propertyKey = NULL;
-	const char * language = NULL;
-	const char * region = NULL;
-	const char * variant = NULL;
-	const char *strings[PROPERTY_COUNT] = {0};
-#define USERNAME_LENGTH 128
-	char username[USERNAME_LENGTH] = {0};
-	char *usernameAlloc = NULL;
-	/* buffer to hold the size of the maximum direct byte buffer allocations */
-	char maxDirectMemBuff[24] = {0};
-	IDATA result = 0;
-
-	J9VMThread *currentThread = (J9VMThread*)env;
-	J9JavaVM *javaVM = currentThread->javaVM;
-	J9InternalVMFunctions *vmFuncs = javaVM->internalVMFunctions;
-
-	/* Change the allocation value PROPERTY_COUNT above as you add/remove properties,
-	 * then follow the propIndex++ convention and consume 2 * slots for each property. 2 * number of property keys is the
-	 * correct allocation.
-	 * Also note the call to addSystemProperties below, which may add some configuration-specific properties.  Be sure to leave
-	 * enough room in the property list for all possibilities.
-	 */
-
-	if (J9_GC_POLICY_METRONOME == (javaVM->omrVM->gcPolicy)) {
-		strings[propIndex++] = "com.ibm.jvm.realtime";
-		strings[propIndex++] = "soft";
-	}
-
-#if defined(J9VM_OPT_SHARED_CLASSES)
-	strings[propIndex++] = "com.ibm.oti.shared.enabled";
-	if ((NULL != javaVM->sharedClassConfig)
-		&& J9_ARE_ALL_BITS_SET(javaVM->sharedClassConfig->runtimeFlags, J9SHR_RUNTIMEFLAG_ENABLE_CACHE_NON_BOOT_CLASSES)
-	) {
-		strings[propIndex++] = "true";
-	} else {
-		strings[propIndex++] = "false";
-	}
-#endif
-
-#if defined(JCL_J2SE)
-	strings[propIndex++] = "ibm.signalhandling.sigchain";
-	if (J9_ARE_ANY_BITS_SET(javaVM->sigFlags, J9_SIG_NO_SIG_CHAIN)) {
-		strings[propIndex++] = "false";
-	} else {
-		strings[propIndex++] = "true";
-	}
-	strings[propIndex++] = "ibm.signalhandling.sigint";
-	if (J9_ARE_ANY_BITS_SET(javaVM->sigFlags, J9_SIG_NO_SIG_INT)) {
-		strings[propIndex++] = "false";
-	} else {
-		strings[propIndex++] = "true";
-	}
-
-	/* The JCLs use ibm.signalhandling.rs to determine if they should prevent the registration of signal handlers for what
-	 * 	we consider to be asynchronous signals.
-	 * The JCLs do not install handlers for any synchronous signals */
-	strings[propIndex++] = "ibm.signalhandling.rs";
-	if (J9_ARE_ALL_BITS_SET(javaVM->sigFlags, J9_SIG_XRS_ASYNC)) {
-		strings[propIndex++] = "true";
-	} else {
-		strings[propIndex++] = "false";
-	}
-#endif
-
-	strings[propIndex++] = "com.ibm.vm.bitmode";
-#ifdef J9VM_ENV_DATA64
-	strings[propIndex++] = "64";
-#else
-	strings[propIndex++] = "32";
-#endif
-
-	strings[propIndex++] = "com.ibm.cpu.endian";
-#ifdef J9VM_ENV_LITTLE_ENDIAN
-	strings[propIndex++] = "little";
-#else
-	strings[propIndex++] = "big";
-#endif
-
-	strings[propIndex++] = "sun.cpu.endian";
-#ifdef J9VM_ENV_LITTLE_ENDIAN
-	strings[propIndex++] = "little";
-#else
-	strings[propIndex++] = "big";
-#endif
-
-/*	Don't set this property as the class library will look here first and when
-	there is a security manager you will get a security exception. The code
-	looks in this package by default, see URLConnection.getContentHandler()
-	strings[propIndex++] = "java.content.handler.pkgs";
-	strings[propIndex++] = "com.ibm.oti.www.content";
-*/
-
-	/*[PR 95709]*/
-
-	/* Get the language, region and variant */
-	language = j9nls_get_language();
-	region = j9nls_get_region();
-	variant = j9nls_get_variant();
-
-	/* CMVC 144405 : Norwegian Bokmal and Nynorsk need special consideration */
-	if ((0 == strcmp(language, "nn")) && (0 == strcmp(region, "NO"))) {
-		variant = "NY";
-	}
-	if ((0 == strcmp(language, "nn")) || (0 == strcmp(language, "nb"))) {
-		language = "no";
-	}
-
-	strings[propIndex++] = "user.language";
-	strings[propIndex++] = language;
-
-	propertyKey = "user.country";
-	strings[propIndex++] = propertyKey;
-	strings[propIndex++] = region;
-
-	/* Get the variant */
-	strings[propIndex++] = "user.variant";
-	strings[propIndex++] = variant;
-
-	/* Get the User name */
-	strings[propIndex++] = "user.name";
-	strings[propIndex] = "unknown";
-#if defined(J9VM_OPT_CRIU_SUPPORT)
-	/* Skip j9sysinfo_get_username if a checkpoint can be taken.
-	 * https://github.com/eclipse-openj9/openj9/issues/15800
-	 */
-	result = -1;
-	if (!vmFuncs->isCheckpointAllowed(javaVM))
-#endif /* defined(J9VM_OPT_CRIU_SUPPORT) */
-	{
-		result = j9sysinfo_get_username(username, USERNAME_LENGTH);
-		if (0 == result) {
-			strings[propIndex] = username;
-		} else if (result > 0) {
-			usernameAlloc = jclmem_allocate_memory(env, result);
-			if (NULL != usernameAlloc) {
-				result = j9sysinfo_get_username(usernameAlloc, result);
-				if (0 == result) {
-					strings[propIndex] = usernameAlloc;
-				} else {
-					/* free the memory, try j9sysinfo_get_env later */
-					jclmem_free_memory(env, usernameAlloc);
-					usernameAlloc = NULL;
-				}
-			} else {
-				vmFuncs->setNativeOutOfMemoryError(currentThread, 0, 0);
-				goto failed;
-			}
-		}
-	}
-#if defined(LINUX) || defined(OSX)
-	if (0 != result) {
-		result = j9sysinfo_get_env("USER", username, USERNAME_LENGTH);
-		if (0 == result) {
-			strings[propIndex] = username;
-		} else if (result > 0) {
-			usernameAlloc = jclmem_allocate_memory(env, result);
-			if (NULL != usernameAlloc) {
-				result = j9sysinfo_get_env("USER", usernameAlloc, result);
-				if (0 == result) {
-					if (strlen(usernameAlloc) > 0) {
-						strings[propIndex] = usernameAlloc;
-					}
-					/* keep it as "unknown" if the env value is empty */
-				}
-				/* usernameAlloc to be freed before this method returns */
-			} else {
-				vmFuncs->setNativeOutOfMemoryError(currentThread, 0, 0);
-				goto failed;
-			}
-		}
-	}
-#endif /* defined(LINUX) || defined(OSX) */
-	propIndex += 1;
-#undef USERNAME_LENGTH
-
-#if defined(OPENJ9_BUILD) && JAVA_SPEC_VERSION == 8
-	/* Set the maximum direct byte buffer allocation property if it has not been set manually */
-	if ((~(UDATA)0) == javaVM->directByteBufferMemoryMax) {
-		UDATA heapSize = javaVM->memoryManagerFunctions->j9gc_get_maximum_heap_size(javaVM);
-		/* allow up to 7/8 of the heap to be direct byte buffers */
-		javaVM->directByteBufferMemoryMax = heapSize - (heapSize / 8);
-	}
-#endif /* defined(OPENJ9_BUILD) && JAVA_SPEC_VERSION == 8 */
-#if !defined(OPENJ9_BUILD)
-	/* Don't set a default value for IBM Java 8. */
-	if ((~(UDATA)0) != javaVM->directByteBufferMemoryMax)
-#endif /* !defined(OPENJ9_BUILD) */
-	{
-		strings[propIndex] = "sun.nio.MaxDirectMemorySize";
-		propIndex += 1;
-		if ((~(UDATA)0) == javaVM->directByteBufferMemoryMax) {
-			strcpy(maxDirectMemBuff, "-1");
-		} else {
-			j9str_printf(maxDirectMemBuff, sizeof(maxDirectMemBuff), "%zu", javaVM->directByteBufferMemoryMax);
-		}
-		strings[propIndex] = maxDirectMemBuff;
-		propIndex += 1;
-	}
-
-	propertyList = getPlatformPropertyList(env, strings, propIndex);
-
-failed:
-	if (NULL != usernameAlloc) {
-		jclmem_free_memory(env, usernameAlloc);
-	}
-	return propertyList;
-}
-
-static void JNICALL
-systemPropertyIterator(char* key, char* value, void* userData)
-{
-	CreateSystemPropertiesData * iteratorData = userData;
-	jobject args = iteratorData->args;
-	JNIEnv *env = iteratorData->env;
-	const char **defaultValues = iteratorData->defaultValues;
-	int defaultCount = iteratorData->defaultCount;
-	jint i;
-
-	/* CMVC 95717: if an error has already occurred get out of here */
-	if ( iteratorData->errorOccurred ) {
-		return;
-	}
-
-	if (0 == strcmp("com.ibm.oti.shared.enabled", key)) {
-		/* JAZZ103 85641: Prevent com.ibm.oti.shared.enabled from being overwritten by a command line option */
-		return;
-	}
-
-#if JAVA_SPEC_VERSION >= 21
-	if (0 == strcmp("java.compiler", key)) {
-		PORT_ACCESS_FROM_ENV(env);
-		if ((0 == strcmp("jitc", value)) || (0 == strcmp(J9_JIT_DLL_NAME, value))) {
-			j9nls_printf(PORTLIB, J9NLS_WARNING, J9NLS_JCL_JAVA_COMPILER_WARNING_XJIT);
-		} else {
-			j9nls_printf(PORTLIB, J9NLS_WARNING, J9NLS_JCL_JAVA_COMPILER_WARNING_XINT);
-		}
-		return;
-	}
-#endif /* JAVA_SPEC_VERSION >= 21 */
-
-	/* check for overridden system properties, use linear scan for now */
-	for (i=0; i < defaultCount; i+=2) {
-		if (defaultValues[i] && !strcmp(key, defaultValues[i])) {
-			defaultValues[i] = NULL;
-			defaultValues[i+1] = NULL;
-			break;
-		}
-	}
-
-	/* First do the key */
-	if( propertyListAddString( env, args, iteratorData->nCommandLineDefines++, key) ) {
-		iteratorData->errorOccurred = 1;
-		return;
-	}
-
-	/* Then the value */
-	if( propertyListAddString( env, args, iteratorData->nCommandLineDefines++, value) ) {
-		iteratorData->errorOccurred = 1;
-		return;
-	}
-
-	Trc_JCL_systemPropertyIterator(env, key, value);
-}
-
 
 void JNICALL
 Java_java_lang_System_startSNMPAgent(JNIEnv *env, jclass jlClass)
