+++ runtime/oti/j9nonbuilder.h
@@ -5350,6 +5350,7 @@ typedef struct J9InternalVMFunctions {
 	struct J9ObjectMonitor * (*monitorTablePeek)(struct J9JavaVM *vm, j9object_t object);
 	jobject (*takeVirtualThreadListToUnblock)(struct J9VMThread *currentThread, struct J9JavaVM *vm);
 #endif /* JAVA_SPEC_VERSION >= 24 */
+	jobjectArray (*getSystemPropertyList)(JNIEnv *env);
 } J9InternalVMFunctions;
 
 /* Jazz 99339: define a new structure to replace JavaVM so as to pass J9NativeLibrary to JVMTIEnv  */
