+++ runtime/vm/vmprops.c
@@ -20,6 +20,13 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0 OR GPL-2.0-only WITH OpenJDK-assembly-exception-1.0
  *******************************************************************************/
 
+#if defined(WIN32)
+#include <windows.h>
+#else /* defined(WIN32) */
+#include <pwd.h>
+#include <unistd.h>
+#endif /* defined(WIN32) */
+
 #include <string.h>
 #include "j9protos.h"
 #include "j9port.h"
@@ -32,24 +39,28 @@
 #include "vmargs_api.h"
 #include "vendor_version.h"
 
+#if defined(J9ZOS390)
+#include "atoe.h"
+#endif /* defined(J9ZOS390) */
+
 #if defined(LINUX)
 /* Copy the system properties names and values into malloced memory */
 static void copySystemProperties(J9JavaVM* vm);
 #endif /* defined(LINUX) */
 
-static U_8*  unicodeEscapeStringToMUtf8(J9JavaVM * vm, const char* escapeString, UDATA escapeLength);
+static U_8 *unicodeEscapeStringToMUtf8(J9JavaVM *vm, const char *escapeString, UDATA escapeLength);
 static UDATA getLibSubDir(J9JavaVM *VM, const char *subDir, char **value);
 
 #define JAVA_ENDORSED_DIRS "java.endorsed.dirs"
 #define JAVA_EXT_DIRS "java.ext.dirs"
 
-UDATA addSystemProperty(J9JavaVM * vm, const char* propName,  const char* propValue, UDATA flags);
-static char * getOptionArg(J9JavaVM *vm, IDATA argIndex, UDATA optionNameLen);
+UDATA addSystemProperty(J9JavaVM *vm, const char *propName,  const char *propValue, UDATA flags);
+static char *getOptionArg(J9JavaVM *vm, IDATA argIndex, UDATA optionNameLen);
 static UDATA addPropertyForOptionWithPathArg(J9JavaVM *vm, const char *optionName, UDATA optionNameLen, const char *propName);
 static UDATA addPropertyForOptionWithModuleListArg(J9JavaVM *vm, const char *optionName, IDATA optionNameLen, const char *propName);
 static UDATA addPropertiesForOptionWithAssignArg(J9JavaVM *vm, const char *optionName, UDATA optionNameLen, const char *basePropName, UDATA basePropNameLen, UDATA *propertyCount);
 static UDATA addPropertyForOptionWithEqualsArg(J9JavaVM *vm, const char *optionName, UDATA optionNameLen, const char *propName);
-static UDATA addModularitySystemProperties(J9JavaVM * vm);
+static UDATA addModularitySystemProperties(J9JavaVM *vm);
 
 /*
  * Create a copy of the given string in allocated memory.
@@ -168,7 +179,7 @@ getOptionArg(J9JavaVM *vm, IDATA argIndex, UDATA optionNameLen)
 {
 	char *option = NULL;
 	char *optionArg = NULL;
-	J9VMInitArgs* j9vm_args = vm->vmArgsArray;
+	J9VMInitArgs *j9vm_args = vm->vmArgsArray;
 
 	if ((argIndex < 0) || ((UDATA)argIndex >= vm->vmArgsArray->nOptions)) {
 		goto _end;
@@ -315,7 +326,7 @@ addPropertyForOptionWithModuleListArg(J9JavaVM *vm, const char *optionName, IDAT
 {
 	IDATA argIndex = -1;
 	UDATA rc = J9SYSPROP_ERROR_NONE;
-	J9VMInitArgs* j9vm_args	= vm->vmArgsArray;
+	J9VMInitArgs *j9vm_args = vm->vmArgsArray;
 	PORT_ACCESS_FROM_JAVAVM(vm);
 
 	argIndex = FIND_ARG_IN_VMARGS_FORWARD(OPTIONAL_LIST_MATCH_USING_EQUALS, optionName, NULL);
@@ -461,7 +472,7 @@ _end:
  * @return returns J9SYSPROP_ERROR_NONE on success, any other J9SYSPROP_ERROR code on failure
  */
 static UDATA
-addModularitySystemProperties(J9JavaVM * vm)
+addModularitySystemProperties(J9JavaVM *vm)
 {
 	UDATA rc = J9SYSPROP_ERROR_NONE;
 
@@ -559,7 +570,7 @@ getLibSubDir(J9JavaVM *vm, const char *subDir, char **value)
 
 	rc = getSystemProperty(vm, "java.home", &javaHomeProp);
 	if (J9SYSPROP_ERROR_NONE == rc) {
-		char* libSubDir = j9mem_allocate_memory(strlen(javaHomeProp->value) + LITERAL_STRLEN(DIR_SEPARATOR_STR) + LITERAL_STRLEN("lib") + LITERAL_STRLEN(DIR_SEPARATOR_STR) + strlen(subDir) + 1, OMRMEM_CATEGORY_VM);
+		char *libSubDir = j9mem_allocate_memory(strlen(javaHomeProp->value) + LITERAL_STRLEN(DIR_SEPARATOR_STR) + LITERAL_STRLEN("lib") + LITERAL_STRLEN(DIR_SEPARATOR_STR) + strlen(subDir) + 1, OMRMEM_CATEGORY_VM);
 		if (NULL == libSubDir) {
 			rc = J9SYSPROP_ERROR_OUT_OF_MEMORY;
 			goto _end;
@@ -586,7 +597,7 @@ _end:
  * @return On success J9SYSPROP_ERROR_NONE, otherwise a J9SYSPROP_ERR constant.
  */
 UDATA
-initializeSystemProperties(J9JavaVM * vm)
+initializeSystemProperties(J9JavaVM *vm)
 {
 	PORT_ACCESS_FROM_JAVAVM(vm);
 
@@ -595,22 +606,22 @@ initializeSystemProperties(J9JavaVM * vm)
 	JavaVMInitArgs *initArgs = NULL;
 	char *jclName = J9_JAVA_SE_DLL_NAME;
 	UDATA j2seVersion = J2SE_VERSION(vm);
-	const char* propValue = NULL;
+	const char *propValue = NULL;
 	UDATA rc = J9SYSPROP_ERROR_NONE;
 	const char *specificationVersion = NULL;
 	BOOLEAN addManagementModule = FALSE;
 
-	if (omrthread_monitor_init(&(vm->systemPropertiesMutex), 0) != 0) {
+	if (0 != omrthread_monitor_init(&(vm->systemPropertiesMutex), 0)) {
 		return J9SYSPROP_ERROR_OUT_OF_MEMORY;
 	}
 
 	/* Count the number of -D properties and find the JCL config */
 	initArgs = vm->vmArgsArray->actualVMArgs;
 	for (i = 0; i < initArgs->nOptions; ++i) {
-		char * optionString = initArgs->options[i].optionString;
+		char *optionString = initArgs->options[i].optionString;
 		Assert_VM_notNull(optionString);
 
-		if (strncmp("-Xjcl:", optionString, 6) == 0) {
+		if (0 == strncmp("-Xjcl:", optionString, 6)) {
 			jclName = optionString + 6;
 		}
 	}
@@ -621,7 +632,8 @@ initializeSystemProperties(J9JavaVM * vm)
 	}
 
 	/* Allocate the properties pool */
-	if ((vm->systemProperties = pool_new(sizeof(J9VMSystemProperty), 0, 0, 0, J9_GET_CALLSITE(), OMRMEM_CATEGORY_VM, POOL_FOR_PORT(vm->portLibrary))) == NULL) {
+	vm->systemProperties = pool_new(sizeof(J9VMSystemProperty), 0, 0, 0, J9_GET_CALLSITE(), OMRMEM_CATEGORY_VM, POOL_FOR_PORT(vm->portLibrary));
+	if (NULL == vm->systemProperties) {
 		return J9SYSPROP_ERROR_OUT_OF_MEMORY;
 	}
 
@@ -632,7 +644,7 @@ initializeSystemProperties(J9JavaVM * vm)
 		}
 	}
 
-	if (JAVA_SPEC_VERSION == 8) {
+	if (8 == JAVA_SPEC_VERSION) {
 		specificationVersion = "1.8";
 	} else {
 		specificationVersion = JAVA_SPEC_VERSION_STRING;
@@ -656,7 +668,7 @@ initializeSystemProperties(J9JavaVM * vm)
 	}
 	{
 		const char *classVersion = NULL;
-		if (JAVA_SPEC_VERSION == 8) {
+		if (8 == JAVA_SPEC_VERSION) {
 			classVersion = "52.0";
 		} else {
 			classVersion = "55.0"; /* Java 11 */
@@ -746,14 +758,14 @@ initializeSystemProperties(J9JavaVM * vm)
 	if (J9SYSPROP_ERROR_NONE != rc) {
 		goto fail;
 	}
-#endif
+#endif /* defined(J9JDK_EXT_NAME) */
 
 #if defined(J9JDK_EXT_VERSION)
 	rc = addSystemProperty(vm, "jdk.extensions.version", J9JDK_EXT_VERSION, 0);
 	if (J9SYSPROP_ERROR_NONE != rc) {
 		goto fail;
 	}
-#endif
+#endif /* defined(J9JDK_EXT_VERSION) */
 
 #if JAVA_SPEC_VERSION < 21
 	/* Don't know the JIT yet, put in a placeholder and make it writeable for now */
@@ -766,9 +778,9 @@ initializeSystemProperties(J9JavaVM * vm)
 	/* We don't have enough information yet. Put in placeholders. */
 #if defined(J9VM_OPT_SIDECAR) && !defined(WIN32)
 	propValue = "../..";
-#else
+#else /* defined(J9VM_OPT_SIDECAR) && !defined(WIN32) */
 	propValue = "..";
-#endif
+#endif /* defined(J9VM_OPT_SIDECAR) && !defined(WIN32) */
 	rc = addSystemProperty(vm, "java.home", propValue, J9SYSPROP_FLAG_WRITEABLE);
 	if (J9SYSPROP_ERROR_NONE != rc) {
 		goto fail;
@@ -796,12 +808,12 @@ initializeSystemProperties(J9JavaVM * vm)
 	if (J9SYSPROP_ERROR_NONE != rc) {
 		goto fail;
 	}
-#endif
+#endif /* defined(J9VM_OPT_SIDECAR) */
 
 	/* Figure out the path separator by querying port library */
 	{
-		char *pathSep = (char*) j9mem_allocate_memory(2, OMRMEM_CATEGORY_VM);
-		if (pathSep == NULL) {
+		char *pathSep = (char *)j9mem_allocate_memory(2, OMRMEM_CATEGORY_VM);
+		if (NULL == pathSep) {
 			return J9SYSPROP_ERROR_OUT_OF_MEMORY;
 		}
 		pathSep[0] = (char) j9sysinfo_get_classpathSeparator();
@@ -911,16 +923,16 @@ initializeSystemProperties(J9JavaVM * vm)
 	/* Create the -D properties. This may override any of the writeable properties above.
 	    Should the command line override read-only props? */
 	for (i = 0; i < initArgs->nOptions; ++i) {
-		char * optionString = initArgs->options[i].optionString;
+		char *optionString = initArgs->options[i].optionString;
 
-		if (strncmp("-D", optionString, 2) == 0) {
+		if (0 == strncmp("-D", optionString, 2)) {
 			J9VMSystemProperty *currentProp = NULL;
 			char *propNameCopy = NULL;
 			char *propValueCopy = NULL;
 			UDATA propNameLen = 0;
 
 			propValue = strchr(optionString + 2, '=');
-			if (propValue == NULL) {
+			if (NULL == propValue) {
 				propNameLen = strlen(optionString) - 2;
 				propValue = optionString + 2 + propNameLen;
 			} else {
@@ -970,7 +982,7 @@ initializeSystemProperties(J9JavaVM * vm)
 				}
 			}
 
-			if (getSystemProperty(vm, propNameCopy, &currentProp) == J9SYSPROP_ERROR_NONE) {
+			if (J9SYSPROP_ERROR_NONE == getSystemProperty(vm, propNameCopy, &currentProp)) {
 				/* We already have a property of the given name, free the propNameCopy buffer */
 				j9mem_free_memory(propNameCopy);
 
@@ -1024,8 +1036,8 @@ initializeSystemProperties(J9JavaVM * vm)
 		}
 	} else {
 		/* look for java.endorsed.dirs and set if not set */
-		if ( getSystemProperty(vm, JAVA_ENDORSED_DIRS, &javaEndorsedDirsProperty) != J9SYSPROP_ERROR_NONE ) {
-			char *defaultEndorsedDir =  NULL;
+		if (J9SYSPROP_ERROR_NONE != getSystemProperty(vm, JAVA_ENDORSED_DIRS, &javaEndorsedDirsProperty)) {
+			char *defaultEndorsedDir = NULL;
 
 			rc = getLibSubDir(vm, "endorsed", &defaultEndorsedDir);
 			if (NULL != defaultEndorsedDir) {
@@ -1117,14 +1129,14 @@ fail:
 
 
 void
-freeSystemProperties(J9JavaVM * vm)
+freeSystemProperties(J9JavaVM *vm)
 {
 	if (NULL != vm->systemProperties) {
 		PORT_ACCESS_FROM_JAVAVM(vm);
 		pool_state walkState;
 
 		J9VMSystemProperty* property = pool_startDo(vm->systemProperties, &walkState);
-		while (property != NULL) {
+		while (NULL != property) {
 			if (property->flags & J9SYSPROP_FLAG_NAME_ALLOCATED) {
 				j9mem_free_memory(property->name);
 			}
@@ -1145,13 +1157,13 @@ freeSystemProperties(J9JavaVM * vm)
 
 
 UDATA
-getSystemProperty(J9JavaVM * vm, const char * name, J9VMSystemProperty ** propertyPtr)
+getSystemProperty(J9JavaVM *vm, const char *name, J9VMSystemProperty **propertyPtr)
 {
 	pool_state walkState;
 
-	J9VMSystemProperty* property = pool_startDo(vm->systemProperties, &walkState);
-	while (property != NULL) {
-		if (strcmp(property->name, name) == 0) {
+	J9VMSystemProperty *property = pool_startDo(vm->systemProperties, &walkState);
+	while (NULL != property) {
+		if (0 == strcmp(property->name, name)) {
 			if (NULL != propertyPtr) {
 				*propertyPtr = property;
 			}
@@ -1174,21 +1186,21 @@ getJ9VMVersionString(J9JavaVM *vm) {
 
 
 UDATA
-addSystemProperty(J9JavaVM * vm, const char* propName,  const char* propValue, UDATA flags)
+addSystemProperty(J9JavaVM *vm, const char *propName, const char *propValue, UDATA flags)
 {
-	J9VMSystemProperty* newProp = pool_newElement(vm->systemProperties);
+	J9VMSystemProperty *newProp = pool_newElement(vm->systemProperties);
 	if (NULL == newProp) {
 		return J9SYSPROP_ERROR_OUT_OF_MEMORY;
 	}
 
-	newProp->name = (char*)propName;
-	newProp->value = (char*)propValue;
+	newProp->name = (char *)propName;
+	newProp->value = (char *)propValue;
 	newProp->flags = flags;
 	return J9SYSPROP_ERROR_NONE;
 }
 
 UDATA
-setSystemPropertyValue(J9JavaVM * vm, J9VMSystemProperty * property, char * newValue, BOOLEAN allocated)
+setSystemPropertyValue(J9JavaVM *vm, J9VMSystemProperty *property, char *newValue, BOOLEAN allocated)
 {
 	/* Make sure the property is writeable */
 
@@ -1210,7 +1222,7 @@ setSystemPropertyValue(J9JavaVM * vm, J9VMSystemProperty * property, char * newV
 			property->flags |= J9SYSPROP_FLAG_VALUE_ALLOCATED;
 		}
 		property->value = newValue;
-		if (strcmp(property->name, "java.home") == 0) {
+		if (0 == strcmp(property->name, "java.home")) {
 			vm->javaHome = (U_8*)newValue;
 		}
 	}
@@ -1218,7 +1230,7 @@ setSystemPropertyValue(J9JavaVM * vm, J9VMSystemProperty * property, char * newV
 }
 
 UDATA
-setSystemProperty(J9JavaVM * vm, J9VMSystemProperty * property, const char * value)
+setSystemProperty(J9JavaVM *vm, J9VMSystemProperty *property, const char *value)
 {
 	/* Make sure the property is writeable */
 
@@ -1230,7 +1242,7 @@ setSystemProperty(J9JavaVM * vm, J9VMSystemProperty * property, const char * val
 
 	if (NULL != value) {
 		/* Make a copy of the value */
-		char * copiedValue = copyToMem(vm, value);
+		char *copiedValue = copyToMem(vm, value);
 
 		if (NULL == copiedValue) {
 			return J9SYSPROP_ERROR_OUT_OF_MEMORY;
@@ -1249,12 +1261,12 @@ setSystemProperty(J9JavaVM * vm, J9VMSystemProperty * property, const char * val
  * @return buffer containing the transliterated string or NULL in case of error
  */
 #define TRANSCODE_BUFFER_SIZE 64
-static U_8*
-unicodeEscapeStringToMUtf8(J9JavaVM * vm, const char* escapeString, UDATA escapeLength) {
+static U_8 *
+unicodeEscapeStringToMUtf8(J9JavaVM *vm, const char *escapeString, UDATA escapeLength) {
 	U_16 localUnicodeBuffer[TRANSCODE_BUFFER_SIZE]; /* handle short strings without allocating memory */
-	U_16* unicodeBuffer = localUnicodeBuffer;
-	UDATA bufferLength = (escapeLength+1)*2; /* This is an overestimate */
-	const char *cursor;
+	U_16 *unicodeBuffer = localUnicodeBuffer;
+	UDATA bufferLength = (escapeLength + 1) * 2; /* This is an overestimate */
+	const char *cursor = NULL;
 	const char *escapeStringEnd = escapeString + escapeLength;
 	IDATA unicodeDigitCounter = -1; /* for walking through "\\u1234" */
 	U_16 currentChar = 0;
@@ -1267,7 +1279,7 @@ unicodeEscapeStringToMUtf8(J9JavaVM * vm, const char* escapeString, UDATA escape
 	PORT_ACCESS_FROM_JAVAVM(vm);
 
 	if (bufferLength > TRANSCODE_BUFFER_SIZE) {
-		unicodeBuffer = (U_16*) j9mem_allocate_memory(bufferLength, OMRMEM_CATEGORY_VM);
+		unicodeBuffer = (U_16 *)j9mem_allocate_memory(bufferLength, OMRMEM_CATEGORY_VM);
 		if (NULL == unicodeBuffer) {
 			return NULL;
 		}
@@ -1275,7 +1287,7 @@ unicodeEscapeStringToMUtf8(J9JavaVM * vm, const char* escapeString, UDATA escape
 
 	cursor = escapeString;
 	while (cursor < escapeStringEnd) {
-		if (cursor[0]=='\\' && cursor[1]=='u') {
+		if (('\\' == cursor[0]) && ('u' == cursor[1])) {
 			slashChar = cursor[0];
 			unicodeDigitCounter = 0;
 			cursor += 2;
@@ -1317,14 +1329,14 @@ unicodeEscapeStringToMUtf8(J9JavaVM * vm, const char* escapeString, UDATA escape
 		cursor++;
 	}
 
-	mutf8Size = j9str_convert(J9STR_CODE_WIDE, J9STR_CODE_MUTF8, (char*)unicodeBuffer, unicodeCharacterCounter*2,
+	mutf8Size = j9str_convert(J9STR_CODE_WIDE, J9STR_CODE_MUTF8, (char *)unicodeBuffer, unicodeCharacterCounter * 2,
 			NULL, 0); /* get the size of the MUTF8 */
 	if (mutf8Size >= 0) {
 		++mutf8Size; /* leave enough space to null-terminate the string */
 		result = j9mem_allocate_memory(mutf8Size, OMRMEM_CATEGORY_VM); /* allow room for terminating null */
 		if (NULL != result) {
-			mutf8Size = j9str_convert(J9STR_CODE_WIDE, J9STR_CODE_MUTF8, (char*)unicodeBuffer, unicodeCharacterCounter*2,
-					(char*)result, mutf8Size);
+			mutf8Size = j9str_convert(J9STR_CODE_WIDE, J9STR_CODE_MUTF8, (char *)unicodeBuffer, unicodeCharacterCounter * 2,
+					(char *)result, mutf8Size);
 			if (mutf8Size < 0) {
 				j9mem_free_memory(result);
 				result = NULL;
@@ -1361,8 +1373,8 @@ containsBackslashU(const char *userString, UDATA stringLength) {
 	if (stringLength > 0) {
 		U_32 cursor = 0;
 		for (cursor = 0; cursor < (stringLength-1); cursor++) {
-			if (('\\' == userString[cursor]) && ('u' == userString[cursor+1])) {
-				/* cursor+1 is safe because the loop stops one character before the end */
+			if (('\\' == userString[cursor]) && ('u' == userString[cursor + 1])) {
+				/* cursor + 1 is safe because the loop stops one character before the end */
 				return TRUE;
 			}
 		}
@@ -1432,4 +1444,648 @@ getMUtf8String(J9JavaVM *vm, const char *userString, UDATA stringLength)
 		j9mem_free_memory(mutf8Buffer);
 	}
 	return result;
+}
+
+/**
+ * @return JNI_FALSE on success, JNI_TRUE on error
+ */
+static jboolean
+propertyListAddString(JNIEnv *env, jarray array, jint arrayIndex, const char *value)
+{
+	/* String must be well-formed modified UTF-8 */
+	jobject str = (*env)->NewStringUTF(env, value);
+	if (NULL != str) {
+		(*env)->SetObjectArrayElement(env, array, arrayIndex, str);
+	}
+	/* NewStringUTF does not throw an exception, other than OutOfMemory */
+	return (*env)->ExceptionCheck(env);
+}
+
+typedef struct {
+	jboolean errorOccurred;
+	jobject args;
+	jint nCommandLineDefines;
+	JNIEnv *env;
+	const char **defaultValues;
+	int defaultCount;
+	jclass stringClass;
+} CreateSystemPropertiesData;
+
+static void JNICALL
+systemPropertyIterator(char *key, char *value, void *userData)
+{
+	CreateSystemPropertiesData *iteratorData = userData;
+	jobject args = iteratorData->args;
+	JNIEnv *env = iteratorData->env;
+	const char **defaultValues = iteratorData->defaultValues;
+	int defaultCount = iteratorData->defaultCount;
+	jint i = 0;
+
+	/* CMVC 95717: if an error has already occurred get out of here */
+	if (iteratorData->errorOccurred) {
+		return;
+	}
+
+	if (0 == strcmp("com.ibm.oti.shared.enabled", key)) {
+		/* JAZZ103 85641: Prevent com.ibm.oti.shared.enabled from being overwritten by a command line option */
+		return;
+	}
+
+#if JAVA_SPEC_VERSION >= 21
+	if (0 == strcmp("java.compiler", key)) {
+		PORT_ACCESS_FROM_ENV(env);
+		if ((0 == strcmp("jitc", value)) || (0 == strcmp(J9_JIT_DLL_NAME, value))) {
+			j9nls_printf(PORTLIB, J9NLS_WARNING, J9NLS_VM_JAVA_COMPILER_WARNING_XJIT);
+		} else {
+			j9nls_printf(PORTLIB, J9NLS_WARNING, J9NLS_VM_JAVA_COMPILER_WARNING_XINT);
+		}
+		return;
+	}
+#endif /* JAVA_SPEC_VERSION >= 21 */
+
+	/* check for overridden system properties, use linear scan for now */
+	for (i = 0; i < defaultCount; i += 2) {
+		if ((NULL != defaultValues[i]) && (0 == strcmp(key, defaultValues[i]))) {
+			defaultValues[i] = NULL;
+			defaultValues[i + 1] = NULL;
+			break;
+		}
+	}
+
+	/* First do the key */
+	if (propertyListAddString(env, args, iteratorData->nCommandLineDefines++, key)) {
+		iteratorData->errorOccurred = JNI_TRUE;
+		return;
+	}
+
+	/* Then the value */
+	if (propertyListAddString(env, args, iteratorData->nCommandLineDefines++, value)) {
+		iteratorData->errorOccurred = JNI_TRUE;
+		return;
+	}
+
+	Trc_systemPropertyIterator(env, key, value);
+}
+
+jobject
+createSystemPropertyList(JNIEnv *env, const char *defaultValues[], int defaultCount)
+{
+	VMI_ACCESS_FROM_ENV(env);
+
+	jint i = 0;
+	jint nCommandLineDefines = 0;
+	jarray args = NULL;
+	int propertyCount = 0;
+	jclass stringClass = (*env)->FindClass(env, "java/lang/String");
+	if (NULL == stringClass) {
+		return NULL;
+	}
+
+	(*VMI)->CountSystemProperties(VMI, &propertyCount);
+	if (0 != propertyCount) {
+		CreateSystemPropertiesData iteratorData;
+
+		args = (*env)->NewObjectArray(env, defaultCount + (propertyCount * 2), stringClass, NULL);
+		if (NULL == args) {
+			return NULL;
+		}
+
+		iteratorData.errorOccurred = JNI_FALSE;
+		iteratorData.args = args;
+		iteratorData.nCommandLineDefines = nCommandLineDefines;
+		iteratorData.env = env;
+		iteratorData.defaultValues = defaultValues;
+		iteratorData.defaultCount = defaultCount;
+
+		iteratorData.stringClass = stringClass;
+		(*VMI)->IterateSystemProperties(VMI, systemPropertyIterator, &iteratorData);
+		if (iteratorData.errorOccurred) {
+			return NULL;
+		}
+		nCommandLineDefines = iteratorData.nCommandLineDefines;
+	}
+
+	if (NULL == args) {
+		args = (*env)->NewObjectArray(env, defaultCount, stringClass, NULL);
+	}
+	if (NULL == args) {
+		return NULL;
+	}
+
+	for (i = 0; i < defaultCount; ++i) {
+		if (NULL == defaultValues[i]) {
+			continue;
+		}
+		if (propertyListAddString(env, args, nCommandLineDefines, defaultValues[i])) {
+			return NULL;
+		}
+		nCommandLineDefines++;
+	}
+
+	return args;
+}
+
+#if defined(WIN32)
+
+/**
+ * @internal
+ * Converts the Unicode string to UTF8 encoded data in the provided buffer.
+ *
+ * @param[in] portLibrary The port library
+ * @param[in] unicodeString The unicode buffer to convert
+ * @param[in] utf8Buffer The buffer to store the UTF8 encoded bytes into
+ * @param[in] size The size of utf8Buffer
+ *
+ * @return 0 on success, -1 on failure.
+ */
+static I_32
+convertToUTF8(J9PortLibrary *portLibrary, const wchar_t *unicodeString, char *utf8Buffer, UDATA size)
+{
+	PORT_ACCESS_FROM_PORT(portLibrary);
+	if (0 == WideCharToMultiByte(OS_ENCODING_CODE_PAGE, OS_ENCODING_WC_FLAGS, unicodeString, -1, utf8Buffer, (int)size, NULL, NULL)) {
+		j9error_set_last_error(GetLastError(), J9PORT_ERROR_OPFAILED); /* continue */
+		return -1;
+	}
+	return 0;
+}
+
+jobject
+getPlatformPropertyList(JNIEnv *env, const char *strings[], int propIndex)
+{
+	PORT_ACCESS_FROM_ENV(env);
+#if !defined(_WIN32_WINNT_WINBLUE) || (_WIN32_WINNT_MAXVER < _WIN32_WINNT_WINBLUE)
+	OSVERSIONINFO versionInfo;
+#endif /* !defined(_WIN32_WINNT_WINBLUE) || (_WIN32_WINNT_MAXVER < _WIN32_WINNT_WINBLUE) */
+	I_32 envSize = 0;
+	char *envSpace = NULL;
+	char *tempdir = NULL;
+	jobject result = NULL;
+	char userhome[EsMaxPath];
+	wchar_t unicodeTemp[EsMaxPath];
+	int i = 0;
+#if JAVA_SPEC_VERSION < 17
+	char userdir[EsMaxPath];
+#endif /* JAVA_SPEC_VERSION < 17 */
+	wchar_t unicodeHome[EsMaxPath];
+	HANDLE process = 0;
+	HANDLE token = 0;
+	UDATA handle = 0;
+	BOOL (WINAPI *func)(HANDLE hToken, LPWSTR lpProfileDir, LPDWORD lpcchSize);
+
+	/* Hard coded file/path separators and other values */
+
+#if JAVA_SPEC_VERSION < 17
+	strings[propIndex++] = "file.separator";
+	strings[propIndex++] = "\\";
+
+	strings[propIndex++] = "line.separator";
+	strings[propIndex++] = "\r\n";
+
+	/* Get the Temp Dir name */
+	strings[propIndex++] = "java.io.tmpdir";
+	strings[propIndex++] = getTmpDir(env, &tempdir);
+#endif /* JAVA_SPEC_VERSION < 17 */
+
+	strings[propIndex++] = "user.home";
+	i = propIndex;
+	envSize = (I_32)j9sysinfo_get_env("USERPROFILE", NULL, 0);
+	if (-1 != envSize) {
+		envSpace = j9mem_allocate_memory(envSize, OMRMEM_CATEGORY_VM); /* trailing null taken into account */
+		if (NULL == envSpace) {
+			strings[propIndex++] = "\\";
+		} else {
+			j9sysinfo_get_env("USERPROFILE", envSpace, envSize);
+			strings[propIndex++] = envSpace;
+		}
+	}
+#if defined(_WIN32_WINNT_WINBLUE) && (_WIN32_WINNT_MAXVER >= _WIN32_WINNT_WINBLUE)
+	/* dwPlatformId, VER_PLATFORM_WIN32_NT = https://msdn.microsoft.com/en-us/library/windows/desktop/ms724834(v=vs.85).aspx */
+	if ((i == propIndex) && IsWindowsVersionOrGreater(5, 0, 0))
+#else /* defined(_WIN32_WINNT_WINBLUE) && (_WIN32_WINNT_MAXVER >= _WIN32_WINNT_WINBLUE) */
+	versionInfo.dwOSVersionInfoSize = sizeof(OSVERSIONINFO);
+#pragma warning(suppress : 4996) /* ignore warning 'GetVersionExA': was declared deprecated */
+	if ((i == propIndex) && GetVersionEx(&versionInfo) && (VER_PLATFORM_WIN32_NT == versionInfo.dwPlatformId))
+#endif /* defined(_WIN32_WINNT_WINBLUE) && (_WIN32_WINNT_MAXVER >= _WIN32_WINNT_WINBLUE) */
+	{
+		process = GetCurrentProcess();
+		if (OpenProcessToken(process, TOKEN_QUERY, &token)) {
+			envSize = 0;
+			if (0 == j9util_open_system_library("userenv", &handle, TRUE)) {
+				if (i == propIndex) {
+					if (0 == j9sl_lookup_name(handle, "GetUserProfileDirectoryW", (UDATA *)&func, "ZPLP")) {
+						envSize = EsMaxPath;
+						if (func(token, unicodeHome, &envSize)) {
+							/* When the SystemDrive environment variable isn't set, such as when j9 is exec'ed
+							 * running JCK tests, we get %SystemDrive%/Documents and Settings/...
+							 */
+							if (0 == wcsncmp(unicodeHome, L"%SystemDrive%", 13)) {
+								/* Borrow userdir variable, which is used for real below */
+								if (GetSystemDirectoryW(unicodeTemp, EsMaxPath) > 1) {
+									unicodeHome[0] = unicodeTemp[0];
+									unicodeHome[1] = unicodeTemp[1];
+									wcsncpy(&unicodeHome[2], &unicodeHome[13], envSize - 13);
+								}
+							}
+							convertToUTF8(PORTLIB, unicodeHome, userhome, EsMaxPath);
+							strings[propIndex++] = userhome;
+						}
+					}
+				}
+			}
+		}
+	}
+
+	if (i == propIndex) {
+		/* Fallback to Windows Directory */
+		envSize = (I_32)j9sysinfo_get_env("WINDIR", NULL, 0);
+		if (-1 == envSize) {
+			strings[propIndex++] = "\\";
+		} else {
+			envSpace = j9mem_allocate_memory(envSize, OMRMEM_CATEGORY_VM);	/* trailing null taken into account */
+			if (NULL == envSpace) {
+				strings[propIndex++] = "\\";
+			} else {
+				j9sysinfo_get_env("WINDIR", envSpace, envSize);
+				strings[propIndex++] = envSpace;
+			}
+		}
+	}
+
+#if JAVA_SPEC_VERSION < 17
+	/* Get the directory where the executable was started */
+	strings[propIndex++] = "user.dir";
+	if (0 == GetCurrentDirectoryW(EsMaxPath, unicodeTemp)) {
+		strings[propIndex++] = "\\";
+	} else {
+		convertToUTF8(PORTLIB, unicodeTemp, userdir, EsMaxPath);
+		strings[propIndex++] = userdir;
+	}
+#endif /* JAVA_SPEC_VERSION < 17 */
+
+	if (JAVA_SPEC_VERSION < 12) {
+		/* Get the timezone */
+		strings[propIndex++] = "user.timezone";
+		strings[propIndex++] = "";
+	}
+
+	result = createSystemPropertyList(env, strings, propIndex);
+	j9mem_free_memory(tempdir);
+	j9mem_free_memory(envSpace);
+	return result;
+}
+
+#else /* defined(WIN32) */
+
+jobject
+getPlatformPropertyList(JNIEnv *env, const char *strings[], int propIndex)
+{
+	PORT_ACCESS_FROM_ENV(env);
+	char *charResult = NULL;
+	char *envSpace = NULL;
+	jobject plist = NULL;
+#if JAVA_SPEC_VERSION < 17
+	char userdir[EsMaxPath] = {0};
+#endif /* JAVA_SPEC_VERSION < 17 */
+	char home[EsMaxPath] = {0};
+	char *homeAlloc = NULL;
+	J9VMThread *currentThread = (J9VMThread*)env;
+	J9JavaVM *vm = currentThread->javaVM;
+	J9InternalVMFunctions *vmFuncs = vm->internalVMFunctions;
+
+	/* Hard coded file/path separators and other values */
+
+#if defined(J9ZOS390)
+	if (J2SE_VERSION_FROM_ENV(env)) {
+		strings[propIndex++] = "platform.notASCII";
+		strings[propIndex++] = "true";
+
+		strings[propIndex++] = "os.encoding";
+		strings[propIndex++] = "ISO8859_1";
+	}
+#endif /* defined(J9ZOS390) */
+
+#if JAVA_SPEC_VERSION < 17
+	strings[propIndex++] = "file.separator";
+	strings[propIndex++] = "/";
+
+	strings[propIndex++] = "line.separator";
+	strings[propIndex++] = "\n";
+
+	/* Get the directory where the executable was started */
+	strings[propIndex++] = "user.dir";
+	charResult = getcwd(userdir, EsMaxPath);
+	if (NULL == charResult) {
+		strings[propIndex++] = ".";
+	} else {
+		strings[propIndex++] = charResult;
+	}
+#endif /* JAVA_SPEC_VERSION < 17 */
+
+	strings[propIndex++] = "user.home";
+	charResult = NULL;
+#if defined(J9ZOS390)
+	charResult = getenv("HOME");
+	if (NULL != charResult) {
+		strings[propIndex++] = charResult;
+	} else {
+		uid_t uid = geteuid();
+		if (0 != uid) {
+			struct passwd *userDescription = getpwuid(uid);
+			if (NULL != userDescription) {
+				charResult = userDescription->pw_dir;
+				strings[propIndex++] = charResult;
+			}
+		} else {
+			char *loginID = getlogin();
+			if (NULL != loginID) {
+				struct passwd *userDescription = getpwnam(loginID);
+				if (NULL != userDescription) {
+					charResult = userDescription->pw_dir;
+					strings[propIndex++] = charResult;
+				}
+			}
+		}
+	}
+
+	/* There exist situations where one of the above calls will fail.
+	 * Fall through to the Unix solution for those cases.
+	 */
+#endif /* defined(J9ZOS390) */
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+	/* Skip getpwuid if a checkpoint can be taken.
+	 * https://github.com/eclipse-openj9/openj9/issues/15800
+	 */
+	if (!vmFuncs->isCheckpointAllowed(vm))
+#endif /* defined(J9VM_OPT_CRIU_SUPPORT) */
+	{
+		/*[PR 101939] user.home not set correctly when j9 invoked via execve(x,y,null) */
+		if (NULL == charResult) {
+			struct passwd *pwentry = getpwuid(getuid());
+			if (NULL != pwentry) {
+				charResult = pwentry->pw_dir;
+				strings[propIndex++] = charResult;
+			}
+		}
+	}
+
+	if (NULL == charResult) {
+		IDATA result = j9sysinfo_get_env("HOME", home, sizeof(home));
+		strings[propIndex] = ".";
+		if (0 == result) {
+			if (strlen(home) > 0) {
+				strings[propIndex] = home;
+			}
+		} else if (result > 0) {
+			homeAlloc = j9mem_allocate_memory(result, OMRMEM_CATEGORY_VM);
+			if (NULL != homeAlloc) {
+				result = j9sysinfo_get_env("HOME", homeAlloc, result);
+				if (0 == result) {
+					strings[propIndex] = homeAlloc;
+				}
+			} else {
+				vmFuncs->setNativeOutOfMemoryError(currentThread, 0, 0);
+				goto failed;
+			}
+		}
+		propIndex += 1;
+	}
+
+#if JAVA_SPEC_VERSION < 17
+	/* Get the Temp Dir name */
+	strings[propIndex++] = "java.io.tmpdir";
+	strings[propIndex++] = getTmpDir(env, &envSpace);
+#endif /* JAVA_SPEC_VERSION < 17 */
+
+	if (JAVA_SPEC_VERSION < 12) {
+		/* Get the timezone */
+		strings[propIndex++] = "user.timezone";
+		strings[propIndex++] = "";
+	}
+
+	plist = createSystemPropertyList(env, strings, propIndex);
+	if (NULL != envSpace) {
+		j9mem_free_memory(envSpace);
+	}
+
+failed:
+	if (NULL != homeAlloc) {
+		j9mem_free_memory(homeAlloc);
+	}
+	return plist;
+}
+
+#endif /* defined(WIN32) */
+
+/**
+ * @brief Get the default system properties.
+ *
+ * @return an Object array of system property key value pairs
+ */
+jobjectArray
+getSystemPropertyList(JNIEnv *env)
+{
+	PORT_ACCESS_FROM_ENV(env);
+	int propIndex = 0;
+	jobject propertyList = NULL;
+#define PROPERTY_COUNT 137
+#if JAVA_SPEC_VERSION < 17
+	char *propertyKey = NULL;
+	const char *language = NULL;
+	const char *region = NULL;
+	const char *variant = NULL;
+#endif /* JAVA_SPEC_VERSION < 17 */
+	const char *strings[PROPERTY_COUNT] = {0};
+#define USERNAME_LENGTH 128
+	char username[USERNAME_LENGTH] = {0};
+	char *usernameAlloc = NULL;
+	/* buffer to hold the size of the maximum direct byte buffer allocations */
+	char maxDirectMemBuff[24] = {0};
+	IDATA result = 0;
+
+	J9VMThread *currentThread = (J9VMThread *)env;
+	J9JavaVM *javaVM = currentThread->javaVM;
+	J9InternalVMFunctions *vmFuncs = javaVM->internalVMFunctions;
+
+	/* Change the allocation value PROPERTY_COUNT above as you add/remove properties,
+	 * then follow the propIndex++ convention and consume 2 * slots for each property.
+	 * 2 * number of property keys is the correct allocation. Also note the call to
+	 * addSystemProperties below, which may add some configuration-specific properties.
+	 * Be sure to leave enough room in the property list for all possibilities.
+	 */
+
+	if (J9_GC_POLICY_METRONOME == javaVM->omrVM->gcPolicy) {
+		strings[propIndex++] = "com.ibm.jvm.realtime";
+		strings[propIndex++] = "soft";
+	}
+
+#if defined(J9VM_OPT_SHARED_CLASSES)
+	strings[propIndex++] = "com.ibm.oti.shared.enabled";
+	if ((NULL != javaVM->sharedClassConfig)
+		&& J9_ARE_ALL_BITS_SET(javaVM->sharedClassConfig->runtimeFlags, J9SHR_RUNTIMEFLAG_ENABLE_CACHE_NON_BOOT_CLASSES)
+	) {
+		strings[propIndex++] = "true";
+	} else {
+		strings[propIndex++] = "false";
+	}
+#endif /* defined(J9VM_OPT_SHARED_CLASSES) */
+
+	strings[propIndex++] = "ibm.signalhandling.sigchain";
+	if (J9_ARE_ANY_BITS_SET(javaVM->sigFlags, J9_SIG_NO_SIG_CHAIN)) {
+		strings[propIndex++] = "false";
+	} else {
+		strings[propIndex++] = "true";
+	}
+	strings[propIndex++] = "ibm.signalhandling.sigint";
+	if (J9_ARE_ANY_BITS_SET(javaVM->sigFlags, J9_SIG_NO_SIG_INT)) {
+		strings[propIndex++] = "false";
+	} else {
+		strings[propIndex++] = "true";
+	}
+
+	/* The JCLs use ibm.signalhandling.rs to determine if they should prevent the
+	 * registration of signal handlers for what we consider to be asynchronous signals.
+	 * The JCLs do not install handlers for any synchronous signals.
+	 */
+	strings[propIndex++] = "ibm.signalhandling.rs";
+	if (J9_ARE_ALL_BITS_SET(javaVM->sigFlags, J9_SIG_XRS_ASYNC)) {
+		strings[propIndex++] = "true";
+	} else {
+		strings[propIndex++] = "false";
+	}
+
+	strings[propIndex++] = "com.ibm.vm.bitmode";
+#if defined(J9VM_ENV_DATA64)
+	strings[propIndex++] = "64";
+#else /* defined(J9VM_ENV_DATA64) */
+	strings[propIndex++] = "32";
+#endif /* defined(J9VM_ENV_DATA64) */
+
+	strings[propIndex++] = "com.ibm.cpu.endian";
+#if defined(J9VM_ENV_LITTLE_ENDIAN)
+	strings[propIndex++] = "little";
+#else /* defined(J9VM_ENV_LITTLE_ENDIAN) */
+	strings[propIndex++] = "big";
+#endif /* defined(J9VM_ENV_LITTLE_ENDIAN) */
+
+#if JAVA_SPEC_VERSION < 17
+	strings[propIndex++] = "sun.cpu.endian";
+#if defined(J9VM_ENV_LITTLE_ENDIAN)
+	strings[propIndex++] = "little";
+#else /* defined(J9VM_ENV_LITTLE_ENDIAN) */
+	strings[propIndex++] = "big";
+#endif /* defined(J9VM_ENV_LITTLE_ENDIAN) */
+
+	/* Get the language, region and variant */
+	language = j9nls_get_language();
+	region = j9nls_get_region();
+	variant = j9nls_get_variant();
+
+	/* CMVC 144405 : Norwegian Bokmal and Nynorsk need special consideration */
+	if ((0 == strcmp(language, "nn")) && (0 == strcmp(region, "NO"))) {
+		variant = "NY";
+	}
+	if ((0 == strcmp(language, "nn")) || (0 == strcmp(language, "nb"))) {
+		language = "no";
+	}
+
+	strings[propIndex++] = "user.language";
+	strings[propIndex++] = language;
+
+	propertyKey = "user.country";
+	strings[propIndex++] = propertyKey;
+	strings[propIndex++] = region;
+
+	/* Get the variant */
+	strings[propIndex++] = "user.variant";
+	strings[propIndex++] = variant;
+#endif /* JAVA_SPEC_VERSION < 17 */
+
+	/* Get the User name */
+	strings[propIndex++] = "user.name";
+	strings[propIndex] = "unknown";
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+	/* Skip j9sysinfo_get_username if a checkpoint can be taken.
+	 * https://github.com/eclipse-openj9/openj9/issues/15800
+	 */
+	result = -1;
+	if (!vmFuncs->isCheckpointAllowed(javaVM))
+#endif /* defined(J9VM_OPT_CRIU_SUPPORT) */
+	{
+		result = j9sysinfo_get_username(username, USERNAME_LENGTH);
+		if (0 == result) {
+			strings[propIndex] = username;
+		} else if (result > 0) {
+			usernameAlloc = j9mem_allocate_memory(result, OMRMEM_CATEGORY_VM);
+			if (NULL != usernameAlloc) {
+				result = j9sysinfo_get_username(usernameAlloc, result);
+				if (0 == result) {
+					strings[propIndex] = usernameAlloc;
+				} else {
+					/* free the memory, try j9sysinfo_get_env later */
+					j9mem_free_memory(usernameAlloc);
+					usernameAlloc = NULL;
+				}
+			} else {
+				vmFuncs->setNativeOutOfMemoryError(currentThread, 0, 0);
+				goto failed;
+			}
+		}
+	}
+#if defined(LINUX) || defined(OSX)
+	if (0 != result) {
+		result = j9sysinfo_get_env("USER", username, USERNAME_LENGTH);
+		if (0 == result) {
+			strings[propIndex] = username;
+		} else if (result > 0) {
+			usernameAlloc = j9mem_allocate_memory(result, OMRMEM_CATEGORY_VM);
+			if (NULL != usernameAlloc) {
+				result = j9sysinfo_get_env("USER", usernameAlloc, result);
+				if (0 == result) {
+					if (strlen(usernameAlloc) > 0) {
+						strings[propIndex] = usernameAlloc;
+					}
+					/* keep it as "unknown" if the env value is empty */
+				}
+				/* usernameAlloc to be freed before this method returns */
+			} else {
+				vmFuncs->setNativeOutOfMemoryError(currentThread, 0, 0);
+				goto failed;
+			}
+		}
+	}
+#endif /* defined(LINUX) || defined(OSX) */
+	propIndex += 1;
+#undef USERNAME_LENGTH
+
+#if defined(OPENJ9_BUILD) && (JAVA_SPEC_VERSION == 8)
+	/* Set the maximum direct byte buffer allocation property if it has not been set manually */
+	if ((~(UDATA)0) == javaVM->directByteBufferMemoryMax) {
+		UDATA heapSize = javaVM->memoryManagerFunctions->j9gc_get_maximum_heap_size(javaVM);
+		/* allow up to 7/8 of the heap to be direct byte buffers */
+		javaVM->directByteBufferMemoryMax = heapSize - (heapSize / 8);
+	}
+#endif /* defined(OPENJ9_BUILD) && (JAVA_SPEC_VERSION == 8) */
+#if !defined(OPENJ9_BUILD)
+	/* Don't set a default value for IBM Java 8. */
+	if ((~(UDATA)0) != javaVM->directByteBufferMemoryMax)
+#endif /* !defined(OPENJ9_BUILD) */
+	{
+		strings[propIndex] = "sun.nio.MaxDirectMemorySize";
+		propIndex += 1;
+		if ((~(UDATA)0) == javaVM->directByteBufferMemoryMax) {
+			strcpy(maxDirectMemBuff, "-1");
+		} else {
+			j9str_printf(maxDirectMemBuff, sizeof(maxDirectMemBuff), "%zu", javaVM->directByteBufferMemoryMax);
+		}
+		strings[propIndex] = maxDirectMemBuff;
+		propIndex += 1;
+	}
+
+	propertyList = getPlatformPropertyList(env, strings, propIndex);
+
+failed:
+	if (NULL != usernameAlloc) {
+		j9mem_free_memory(usernameAlloc);
+	}
+	return propertyList;
 }
